requirements.txt
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.2
django-debug-toolbar==2.0
django-heroku==0.3.1
googlemaps==3.1.3
gunicorn==19.6.0
idna==2.8
isort==4.3.20
mysqlclient==1.4.2.post1
Pillow==6.0.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
python-decouple==3.1
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==3.2
